requirements.txt
Flask>=3.0,<4
opencv-python-headless>=4.9,<5
numpy>=1.26,<3
Pillow>=10,<12
